@@ -1,11 +1,11 @@
 {
-  "name": "crna",
+  "name": "combust-mobile",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-native-scripts": "1.11.1",
-    "jest-expo": "25.0.0",
-    "react-test-renderer": "16.2.0"
+    "react-native-scripts": "1.14.0",
+    "jest-expo": "~27.0.0",
+    "react-test-renderer": "16.3.1"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
@@ -13,19 +13,19 @@
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "node node_modules/jest/bin/jest.js"
+    "test": "jest"
   },
   "jest": {
     "preset": "jest-expo"
   },
   "dependencies": {
-    "expo": "^27.0.0",
-    "firebase": "^4.10.1",
-    "mobx": "^3.6.1",
-    "mobx-react": "^4.4.2",
-    "react": "16.2.0",
-    "react-native": "^0.55.0",
-    "react-native-elements": "^0.19.0",
-    "react-navigation": "^1.4.0"
+    "expo": "^27.0.1",
+    "firebase": "^5.0.2",
+    "mobx": "^4.2.1",
+    "mobx-react": "^5.1.2",
+    "react": "16.3.1",
+    "react-native": "~0.55.2",
+    "react-native-elements": "^0.19.1",
+    "react-navigation": "^2.0.1"
   }
 }
